Part 1-3 Demos Complete
Implemented a program to continually increment a variable in RW memory
and display it on LED17-18. SW1 and SW2 are added and the result
displayed on LED19-20 at the same time. Additionally, SW1 and SW2 are
displayed on LED_Red and LED_Blue respectively.

This functionality was sufficient to get Part 1, 2, and 3 lab demos
signed off, so close #5, close #8, and close #11.

#9 is still open, because not all ALU functions have been tested yet.